Data structure and method changes, different solving function
Changed conflict/contradiction vectors to numeric representation and
modified corresponding methods.
Wrote a new version of the solveBoard function to recursively search for
the solution by setting and removing cells.
Created a new helper function (findEmpty) to find the first empty cell
in the matrix, or return false if there are no more empty cells
(implying the board should be solved).
Modified setCell function to place a value in a cell even if there is a
conflict
Added recursive call count for the solving function by placing static
variable in the class and incrementing it within the function (not quite
sure how to best implement that part). It is also reset within the
initialize function.